Initial checkin.
The new client is Couchbase.CouchbaseClient:

    var client = new Couchbase.CouchbaseClient(...);

Parameters are similar to the MembaseClient, except

- the config class should be ICouchbaseClientConfiguration; because of the design time views we need to know about an INameTransformer
- the config section is configuration/couchbase

The client has a new `GetView` operation which returns an ICouchbaseView. The resulting object does not uery the server until it's enumerated, so parameters can be set conveniently:

    var view = client.GetView("doc", "index").Limit(20);

    foreach(var row in view)
    {
        var item = row.GetItem();
    }

Enumerating the view is convenient if we want to get the new values emitted by the mapper or we want to get the keys, but very inefficient if we need only the items, so the client has a new `Get` overload:

    var items = client.Get(view);

This will do a multiget using all keys returned by the view, and store everything in an IDictionary<string, object> (just like the regular multiget).