Removed the Json.NET dependency.
The client only uses a very small subset of the Newtonsoft library, namely the forward only json parser.

Since Json.NET is MIT licensed it was possible to extract only the neccessary parts of the library (see externals/Newtonsoft) and use them internally. This way we can cut back on the exposed types and users will not feel forced to use Json.NET for json de/serialization.