Migrating Membase client to Couchbase namespaces
With the 1.8 CBS release, the .NET Couchbase client is being updated
to include proper Couchbase class and namespace naming.  The new Couchbase
project found here originated as a fork of the Membase client
located at https://github.com/enyim/EnyimMemcached.

Commit 0c6fe754b2da57c72a688a7bf071092e0d7c7c8b was the starting point.
The client in this commit is not yet signed as was the Membase client.

Change-Id: I0c2d2bdf47e5639df727ca870ef22e0320cfe89e
Reviewed-on: http://review.couchbase.org/12023
Tested-by: John C. Zablocki <john@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>