Added script to remove Enyim.Caching assembly signing
Enyim.Caching, which is a submodule, is setup to use
delay signing when the private key is unavailable (which
will always be the case).  This setup means that developers
who clone the Couchbase repository will be have to either
a) modify the project to use a new key pair b) disable assebmly
verification or c) modify the project to remove the signing.

This new script uses method c as there's no reason
to leave signing in place for local development.  It's also
easier than having to disable assembly verification for
the Enyim.Caching assembly.

Change-Id: If66fca0ca0aaf8314d786111da629eb57f677da4
Reviewed-on: http://review.couchbase.org/14215
Tested-by: John C. Zablocki <john@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>