Added tracing functionality for a key or set of keys using a regex
I added a TraceSwitch to the Couchclient class and a KeyFilter which allows a regex or a string literal representing a key to be used as a filter for tracing keys. Since not all TraceEventTypes are supported, I added a TraceSwitchExtensions class with extension methods to simplify things. I also updated each method in CouchbaseClient that takes a key as a parameter to have a TraceSwitch.TraceKey(key) call. Tracing can be enabled in the system.diagnostics section in the app.config and disabled by either removing the section or changing the switch value to Off. Note that only Verbose, Information and Off are supported. Also, DefaultTraceListener and ConsoleTraceListener are currently configured, but TextWriterTraceListener can be added to write the trace messages to disk.