NCBC-289: Does not return errors object on exception
We decided to throw an exception when an error is detected for all view
error cases. In the next version (2.x) of the client will make a decision
on how we want the client to behave when an error is encountered when
processing a view. This commit makes it consistent across all error cases
and does not change the interface, which would likely impact users
requiring them to change there code from handling exceptions to checking
an errors property for failures.

This commit also adds additional unit tests and refactors the
CouchbaseViewHandler class so that we can pass streams into the
ReadResponse method that contain text resembling errors returned from the
server.

Change-Id: I87f982a1b33236eb3b7aa47eb08a4ef88892500e
Reviewed-on: http://review.couchbase.org/29173
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>