NCBC-309: Move .NET API documentation from docs repo to auto-doc
This is the first step in the process - adding the xml comments to each method
declaration of the publc API. These will be used to generate the HTML API
reference. Note, that in a MVP fashion, I did the bare minimum documentation
for each method. In future releases, the documentation should be improved

Change-Id: I7ce1a5cf2f1e286c98f1925fd8cee5b0b8e68dcf
Reviewed-on: http://review.couchbase.org/29859
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>