NCBC-299: Update Project References from Nuget to Local Project.
In the .NET solution in Github, the some of the projects are referencing
Nuget instead of the local projects within the solution. As a result,
when you pull the repo down you cannot step into the dependent code and
in some cases you need to disable features (such as logging) or you will
get a version mismatch error.

Change-Id: Ic9dc6bb82fd3fefc62f9b645ee1fc8a4a05bf162
Reviewed-on: http://review.couchbase.org/29990
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>