NCBC-311: Change queueTimeout default from 100ms to 2500ms
The default value for queueTimeout is currently 100ms, which is extremely
low and will unnecessarily cause queue timeout exceptions. This will increase
the queueTimeout to 2.5 seconds, which is still a relatively low amount. Note
that this only affects threads waiting on the SocketPool for a socket, not the
actual time it takes to execute an operation.

Change-Id: I43fb2368d92e936cb9d15db8b6ba48df46f13413
Reviewed-on: http://review.couchbase.org/30275
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>