NCBC-337: Fix for 'View vquery was mapped to a dead node, failing.' errors
When a config update occurs, the client may receive a configuration with
nodes with a 'status' of 'warmup', in this state the 'couchbaseApiBase' is not
returned. This puts the client in a state where a client node may not have a
IHttpClient to execute the view request again, thus the request fails. This
commit adds a check to ensure that the node is 'valid' by checking for nodes
where the IHttpClient is not null

Change-Id: If61b463f5458a26d953b65e4391cb209cb9e0e46
Reviewed-on: http://review.couchbase.org/31007
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>