NCBC-352: Flag all Increment/Decrement methods with CAS params as 'Obsolete'
Some of the Increment/Decrement methods in the Enyim libraries have parameters
for CAS values, this commit flags those methods as obsolete and they may be
removed in future versions of the .NET SDK. The reason that they are being
flagged as obsolete is because Increment/Decrement are atomic operations on the
server, thus making the CAS parameter redundent. Furthermore, testing indicates
that they do not work as expected.

Change-Id: I7a0db7891abcbbbaf7463e123c65bdb92c24c821
Reviewed-on: http://review.couchbase.org/31219
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>