NCBC-334: Add a method of getting the version of the library programatically
This commit adds a git post-merge hook that will run after a pull, assuming that the remote
repo has changes. It will use git-describe to get the lastest revision since the last tag
and includes the current SHA1. It will then update a Version.txt file and the AssemblyInfo.cs
class with this information in the AssemblyInformationalVersion attribute. A future enhancement
will update the AssemblyVersion and AssemblyFileVersion as well as add the git log (release notes)
for the current release as an embedded resource and expose a public method to access it.

Change-Id: If969393b134f37555f33c057d6b97f5ba1472713
Reviewed-on: http://review.couchbase.org/31597
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>