NCBC-358: Sort nodes to reduce number server config changes
When a server config update is signaled, the GetHashCode() method compares
the new config and the previous config to determine whether or not to recreate
the connection pool et al because the cluster has changed on the server. In
some cases the only change is to the order of the nodes. This fix orders the
node list before comparing so that the only 'real' changes result in a conn
pool reconfigure.

Change-Id: Ia19c966efce7547057265fd8220d92cd0efea7b1
Reviewed-on: http://review.couchbase.org/31834
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>