NCBC-368: Randomize Bootstrap Node
Randomizes the selection of the bootstrap node so that all clients do not
use the same node, which can cause "stampeding herd" and lopsided resource
utilization in large clusters or when cluster supports a large number of
clients. This ensures that boot-strapping is evenly distributed over all of the
urls in the configured boot-strap list. Note that the first 'good' node is
chosen as a bootstrapper.

Change-Id: I9755367a0a71c6b60cc3f3a913e44f49762f168f
Reviewed-on: http://review.couchbase.org/32379
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>