NCBC-408: Retry Node Check during View Operations
During a View operation, the client will look for a node to use to execute
the operation against. In certain cases (rebalance, etc) the list of nodes
may be empty. This will add an additional check if a candidate node is not
found before the client gives up and returns a "View {0} was mapped to a
dead node, failing." error.

Change-Id: Iecfb45855e9e8c113eb7f8424fd039074a7c6f0f
Reviewed-on: http://review.couchbase.org/34123
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>