NCBC-416: Method overloading ambiguity with named parameters
This is a fix for a regression bug where the method overloading was
differentiated with return type only; hence wrong method was invoked from
public IView GetView(string designName, string viewName). This caused
the rows to be null when iterating over a view with no output other than
it's keys.

Change-Id: Ief3a66588e2b8a30bdbb7038366fc8904f76bf82
Reviewed-on: http://review.couchbase.org/35108
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>