NCBC-553: Could not bootstrap bucket with ssl connection
Fixes configuration bug that caused the bucket configuration settings to
be ignored. This in turn would make the client fail when attempting to
bootstrap because the urls and ports used would be non-ssl. Additional
unit tests and a property (IsSecure) were added to make it easier to
determine if the connection is indeed secure.
Change-Id: Icb174a516cf3e529154d5b51ccd23bcf1f77cc57