NCBC-592: AuthenticationException in CouchbaseBucket.SendWithRetry
This commit fixes the case that a non-auth related exception that
happens during authentication, does not crash the client and is instead
handled as generic operation failure. There are two cases in which an
AuthenticationException will occur now: when the username or password is
not valid for a given bucket or when the bucket requested does not exist.
Note that in the case of non-existent bucket, the server will return an
authentication error.

Change-Id: Ibfec91fe9a44ee47e51d9cabe455e34c358bc16f
Reviewed-on: http://review.couchbase.org/40772
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>