Fixed PagedView with only one page
When using a PageView with a page size larger than the number of items
in the result set, the MoveNext() operation indicates there are no items
in the current page and returns false.

The fix changes the expression to check if there the nextId is not null
and if the current page contains any items.