NCBC-616: Fixed PagedView with only one page
When using a PageView with a page size larger than the number of items
in the result set, the MoveNext() operation indicates there are no items
in the current page and returns false.

The fix changes the expression to check if there the nextId is not null
and if the current page contains any items.

Change-Id: I1b11f3820606f77e937d839ec7e7de48d4dfa6af

Initial load of paged view checks item collection to determine if items are available

Change-Id: I9a19b728945aa3ac639034efe779eb0b3789fe70
Reviewed-on: http://review.couchbase.org/41122
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>