NCBC-662: Ensure Config is decoded as an object
During a rebalance, a operation may result in a NMV with the body of the
response the new config. The response will not contain any common flag
information, since it's not response of the operation, but the response of
a NMV. If the anticipated type is not an object, a cast exception is
thrown. This patch ensures that all NMV payloads are decoded and
deserialized as BucketConfig object.

Change-Id: Icd7a3d40dc8e3ec4c955e87b50df564c281b1d8a
Reviewed-on: http://review.couchbase.org/41815
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>