NCBC-683: When a node goes on/offline ensure that the cluster map is up to date
When a node crashes or otherwise goes offline, the client will continue to
use the stale configuration. If the node becomes active again, the client
will be in an inconsistent state because the node will be flagged as dead
and unusable. This commit ensures that the client gets a fresh
configuration in the case of a NMV or certain IO errors (forcibly closed
or reset) even if the revision hasn't changed.

Change-Id: Id0cd177d38027e57649ddbfd50a00a0c1bf89ac6
Reviewed-on: http://review.couchbase.org/42101
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>