NCBC-696: Use default settings for SecurityProtocol
The ServicePointManager.SecurityProtocol controls the security protocol
that web clients will use if they are a secured channel (SSL). By default
the settings for these are SecurityProtocol.Tls | SecurityProtocol.Ssl3.

There is an exploit of the Ssl3 protocol known as the "POODLE Attack", so
it's advised that the users of the client (or anyone using SSL) use TLS
instead of SSL3. This commit removes the explicit setting of
ServicePointManager.SecurityProtocol in the client, so that the default is
used. Users of the client can and should override this and set it to TLS
in their application servers.

Note that POODLE is very difficult to exploit from an SDK client, but in
general removing SSL3 is good to avoid future exploits of this no-longer-
recommended protocol.

Change-Id: I7dd49cefc85ebdad151deabbe65b87d0bf47a703
Reviewed-on: http://review.couchbase.org/42395
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>