NCBC-691: Default lock time to 15 seconds
The maximum lock in 2.5.1 and 3.0.0 is 30 secs, the default is 15 seconds.
Any attempt to set a lock higher than the maximum results in the default
being used. This patch sets the default locktime to 15 seconds, which was
previously TimeSpan.Zero. The code already throws a AOOR exception if the
locktime is longer than 30 seconds.

Change-Id: I6d1eddf6b670b4afc6340931d78a4fc7b9750b7b
Reviewed-on: http://review.couchbase.org/42675
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>