NCBC-722: Add support for mapping view results to POCOs
This commit changes the way view results are mapped to POCOs via the
ViewRow object. It also allows a consumer to call a Values property to get
the values of the Rows without having to deal with the RowView object.
The result of the Values property is a mapping of the Value of each row
and a POCO into an IEnumerable object.

Change-Id: I07a1c0b053673079fca83bf36f1f839e0f1496aa
Reviewed-on: http://review.couchbase.org/43583
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>