Ensure any operation that returns a NMV is retried
Motivation
----------
When an operation fails with a ResponseStatus.VBucketBelongsToAnotherServer
the client should check for the new configuration that is returned as the body
of the response and force a reconfiguration with it, then re-send the
operation until the operation times out.

Modifications
-------------
The CanRetryOperation(..) method has been modified so that it will return
with a true if an NMV response is encountered and the reconfiguration has
taken place.

Result
------
When an operation returns with a NMV response, the client will perform a
reconfiguration and retry the operation regardless of the operation's
type.

Change-Id: I25af817f95e9ddcd7615a7dbfd80050aad0c93c9
Reviewed-on: http://review.couchbase.org/43819
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>