NCBC-728: Rename IOperationResult.Value to IOperationResult.Content
Motivation
----------

The Java SDK 2.0 equivalent type has a getter/setter named Content that
represents the value of the key, or more specifically, the document that
is stored as JSON in Couchbase. This commit changes the .NET equivalent
from Value to Content.

Modifications
-------------
The IDocument and IDocumentResult interfaces 'Value' property has been
changed to 'Content'. All classes and tests that use those types and
properties have been updated as well.

Result
------
The .NET and Java API's have common fields that behave and are named
equivalently. User's that switch between the the API's will find the
interfaces are consistent.

Change-Id: I6100fa07ca8c2c888994cf290a6fa2e9e2244745
Reviewed-on: http://review.couchbase.org/43825
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>