Make OperationBase's creation time field a public property
Motivation
----------

Each operation has a creation datetime used for managing the lifetime of
the operation. Previously, it was a private variable, this commit makes it
a public property so that the value can be copied when an operation is
cloned.

Modifications
-------------
The _creationTime field has been removed and replaced with a CreationTime
property.

Result
------
When an operation is retried, it is cloned first and the CreationTime
property is copied over from the previous object.

Change-Id: I6a16a1d5b9ecaab1be1cc752d2a550668173f2a2
Reviewed-on: http://review.couchbase.org/43859
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>