NCBC-729: Ensure bulk methods use thread-safe dictionaries
Motivation
----------
The bulk methods for Get(List) and Upsert(Dictionary) use the TPL for
concurrency, however the accumulator dictionary which stors the results is
not thread safe.

Modifications
-------------
Each of the Get and Upsert methods that perform bulk operations now
internally store the results in a ConcurrentDictionary.

Result
------
Errors that were related to threading are no longer encountered.

Change-Id: I27e04dafdcda8c6b9b321cb0d80dc49d83d0a972
Reviewed-on: http://review.couchbase.org/43870
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>