Make finalizers conditional debug only
Motivation
----------
In general finalizers are not needed when dealing exclusively with managed
resources, however they are very useful for development/debugging in that
they can log when a resource is 'leaked' - when it's Dispose method is not
explcilty called. This changset makes the finalizers conditional for DEBUG
builds only.

Modifcations
------------
All finalizers have been wrapped with conditional DEBUG statements.

Result
------
No finalizers will be included in release builds.

Change-Id: I85c31458efaf1782c589dfe3966c5bf43e2b38b3
Reviewed-on: http://review.couchbase.org/43917
Reviewed-by: Brett Lawson <brett19@gmail.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>