NCBC-735: SSL overridden only if set globally
Motivation
----------
In the ClientConfiguration, the UseSsl parameter can be overridden on
bucket configuration by the same parameter being set to true at global
level.
However, in the current state of things this override is also performed
when at least one of the bucket configurations declares UseSsl = true.
This is a mistake and should be corrected.

Modifications
-------------
Took the SSL override code one step up in the code, removed loop and
fixed the boolean test : now UseSsl is only forced on bucket configs if
it was set to true at the root of the config.

Added a unit test to check this behavior.

Result
------
Ssl should not be wrongly forced on each bucket if only one of them has
it enabled.
Corresponding unit test passes.

Change-Id: I6ce7afd68d8e4d94f608352cf3f8210e3e34a659
Reviewed-on: http://review.couchbase.org/44147
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>