NCBC-480: Implement Cluster.Info
Motivation
----------
The Cluster.Info interface is yet not implemented. We need a minimum
implementation that can return information on pools and a list of the
bucket configurations.

Modifications
-------------
Added IClusterInfo implementation that can be internally initialized by
a IServerConfig (internal configuration).
This is instanciated by getting said internal configuration from the
Http configuration provider, in ClusterController.
Cluster.Info delegates to the IClusterController.

Added a test to check retrieval of the ClusterInfo.

Result
------
Calling Cluster.Info no longer throws an Exception but returns info
copied from the internal IServerConfiguration : Pools and List of
IBucketInfo.
Test passes.

Change-Id: I87df336db33b9ff808a845969803f95a2f1ac15d
Reviewed-on: http://review.couchbase.org/44179
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>