Update ViewRow.cs
NCBC-722 changed the result of a view so it is returned as ViewRow<T>, which is fine - BUT 
the Key here is returned as object instead of dynamic - making it a hassle to first unbox to dynamic/Jarray to be able to deserialize it / cast it to JValue and then convert. ( which make this unnecessary verbose and cluttered solution ) 
We need to use Key due to mapreduce - so a better solution is to bring it back as dynamic, like it was before.