Change ViewRow.Id from object to dynamic
Motivation
----------
The id of a key returned from a view may be represented as a JSON object.
This is stored in an property of the ViewRow class as a CLR object Type;
not as the JSON string, but as a JSON.NET object which requires "unboxing"
to convert it to usable key. This fix changes the ViewRow.Id property to a
dynamic type which makes it easier to work with the underlying JSON string
without dealing with the JSON.NET Types.

Modifications
-------------
The Id property on RowView has been changed from an object Type to a
dynamic Type.

Result
------
When using the Id of the RowView class, you no longer have to convert the
Id to a JSON.NET Type to get to the value.

Change-Id: Ibd32032b24ed7a6790458666c46e75fd4e903fe4
Reviewed-on: http://review.couchbase.org/45050
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>