NCBC-739: Test handling Casing in Json transcoding
Motivation
----------
We need to ensure and show in a unit test that by default serialized
json will be camelCased whereas it is easily possible to also produce
PascalCased json.

Modifications
-------------
Added a simple set of unit tests to check and demonstrate json handling
of case in various scenarii. See JsonDataMapperTests.cs and also tests
in DefaultTranscoderTests.cs.

DefaultTranscoderTests was renamed from TypeSerializer2Tests.

Result
------
PascalCased json can be obtained by using a DefaultContractResolver in
the SerializationConfiguration. This is checked and demonstrated in the
new test fixture JsonDataMapperTests and in new tests in
DefaultTranscoderTests.

Change-Id: I1ef8edbeef78ce7bf5a27168802488ee0cd974e7
Reviewed-on: http://review.couchbase.org/44355
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>