NCBC-750: Ensure changes with Server are propagated to BucketConfig
Motivation
----------
When the ClientConfiguration.Servers list is modified, changes need to be
propogated to the BucketConfig elements so that the client can bootstrap
properly.

Modifications
-------------
The ClientConfiguration.Initialize method has been augmented to detected
changes in the Servers list in this case where Clear() and Add() have been
used to change the content of the collection.

Result
------
When a user calls ClientConfiguration.Servers.Clear|Add, the changes will
be detected and the BucketCOnfigs updated with the changes.

Change-Id: Ida65ba712fa43a1079c1efe5f3711355692e02a3
Reviewed-on: http://review.couchbase.org/44370
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>