NCBC-748: Ensure PoolConfiguration is used from App.Config
Motivation
----------
This patch fixes a bug where the values of the poolConfiguration element
where overridden in the initialization process of the ClientConfiguration
class by the default PoolConfiguration.

Modifications
-------------
The _poolConfigurationChanged field is set to true in the ctor overload
which takes the CouchbaseClientSection representing the App.Config
configuration.

Result
------
The correct configuration is used to create the connection pool when using
App.Config configuration.

Change-Id: I84d00dea2bb4b39b981dc7206cc3d7dc73ee96b9
Reviewed-on: http://review.couchbase.org/44443
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>