Fix CouchbaseBucketTests.TestMultiRemove so that it passes
Motivation
----------
The test had a ClientCOnfiguration instantiated with a poolsize of
MaxSize=10 and MinSize=5, but it wasn't passing the config to a Cluster
object; it was using the _cluster object defined as a field with a default
ClientConfiguration. TPL (using at least 8 threads on my machine) simply requires a bigger pool size than the
default of 2 connections, so it would eventually timeout operations
waiting for a connection.

Modifications
-------------
Created a Cluster scoped to the method and passed in the custom
configuration with MaxSize=10 and MinSize=5 connections.

Result
------
The test now passes.

Change-Id: Ic175a03e570bc2239c93e794574dd8c884739073
Reviewed-on: http://review.couchbase.org/44449
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>