Fix CarrierPublicationTests which depend upon fakes
Motivation
----------
This fixes a test that was breaking due to a change in how we handle
NMV's. It also improves/fixes the FakeOperation class that had been
neglected after refactoring the real Operation classes.

Modifications
-------------
Changed the name of the test to represent the current behavior of an
operation which received a NMV. Refactored FakeOperationResult to inherit
from OperationResult and implement IOperationresult<T>.

Result
------
The test now passes.

Change-Id: Ida9a79bb64fc59412725cd145ae92bd0ef98a726
Reviewed-on: http://review.couchbase.org/44450
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>